+++ march_ws/src/march_gait_scheduler/src/Scheduler.cpp
@@ -1,6 +1,11 @@
 // Copyright 2019 Project March.
 #include <march_gait_scheduler/Scheduler.h>
 
-void Scheduler::delayTrajectory(trajectory_msgs::JointTrajectory &trajectory, ros::Duration duration) {
-
+void Scheduler::delayGait(march_rqt_gait_generator::MarchGait& marchGait, ros::Duration delay)
+{
+  //  for(int i =0; i < marchGait.joint_trajectory.points.size(); i++){
+  //    trajectory_msgs::JointTrajectoryPoint point = marchGait.joint_trajectory.points[i];
+  //    point.time_from_start+=delay;
+  //    marchGait.joint_trajectory.points[i] = point;
+  //  }
 }
