+++ march_ws/src/march_gait_scheduler/test/simple-test.test
@@ -1,4 +1,4 @@
 <launch>
     <!-- Launch the required nodes here -->
-    <test test-name="unit_tests" pkg="march_gait_scheduler" type="march_gait_scheduler-test"/>
+    <test test-name="march_gait_scheduler_tests" pkg="march_gait_scheduler" type="march_gait_scheduler-test"/>
 </launch>
